code/requirements.txt
streamlit
pandas
numpy
pyodbc
matplotlib
plotly.express
